Make the `string.Join()` method more fluent (#8)
* Fix test folder name

* Correct test folder name - part two

* Add string.join extensions

Added the extensions based on the official `string.Join()` methods.

* Added some tests for the string.Join extensions

* Remove extensions for code that does not work

These versions of `string.Join` return the object as a string, not the collection joined by the separator.

* Update ReadMe with `Join()` extension

* Correct spelling in enum test method

* Tweak the ReadMe for grammar

* WIP - re-work extension

Changed the extension to not use the additional parameters as they get messed up when using the original.

* Move the extension to the string class

MJ prefered the extension be with the `string` extensions even though we are extending the `IEnumerable`.

* Remove extension using index and count

The call to the main `string.Join()` confuses the additional parameters with the source.

* Add extension for generic type

Requires the generic type to provide an override of the `ToString()` method.

* Update the ReadMe with the move and examples

* Move `Join` extension to `IEnumerable` class

Moved the extensions and tests from the `string` class to the `IEnumerable` class. Updated the `ReadMe.md` file.